Centered dragon to the marker, added roar sound on tap
Added an animation when the dragon is made visible in order to move it back a bit and keep it's feet on the marker. When the user taps the screen, initialized a sound object, loaded a .mp3 of Toothless roaring into it, and played that sound object so that Toothless roars every time the user taps the screen.